+++ src/Zone/Signal.lua
@@ -1,109 +1,178 @@
-local HttpService = game:GetService("HttpService")
-local RunService = game:GetService("RunService")
-local heartbeat = RunService.Heartbeat
+--------------------------------------------------------------------------------
+--               Batched Yield-Safe Signal Implementation                     --
+-- This is a Signal class which has effectively identical behavior to a       --
+-- normal RBXScriptSignal, with the only difference being a couple extra      --
+-- stack frames at the bottom of the stack trace when an error is thrown.     --
+-- This implementation caches runner coroutines, so the ability to yield in   --
+-- the signal handlers comes at minimal extra cost over a naive signal        --
+-- implementation that either always or never spawns a thread.                --
+--                                                                            --
+-- API:                                                                       --
+--   local Signal = require(THIS MODULE)                                      --
+--   local sig = Signal.new()                                                 --
+--   local connection = sig:Connect(function(arg1, arg2, ...) ... end)        --
+--   sig:Fire(arg1, arg2, ...)                                                --
+--   connection:Disconnect()                                                  --
+--   sig:DisconnectAll()                                                      --
+--   local arg1, arg2, ... = sig:Wait()                                       --
+--                                                                            --
+-- Licence:                                                                   --
+--   Licenced under the MIT licence.                                          --
+--                                                                            --
+-- Authors:                                                                   --
+--   stravant - July 31st, 2021 - Created the file.                           --
+--------------------------------------------------------------------------------
+
+-- The currently idle thread to run the next handler on
+local freeRunnerThread = nil
+
+-- Function which acquires the currently idle handler runner thread, runs the
+-- function fn on it, and then releases the thread, returning it to being the
+-- currently idle one.
+-- If there was a currently idle runner thread already, that's okay, that old
+-- one will just get thrown and eventually GCed.
+local function acquireRunnerThreadAndCallEventHandler(fn, ...)
+	local acquiredRunnerThread = freeRunnerThread
+	freeRunnerThread = nil
+	fn(...)
+	-- The handler finished running, this runner thread is free again.
+	freeRunnerThread = acquiredRunnerThread
+end
+
+-- Coroutine runner that we create coroutines of. The coroutine can be 
+-- repeatedly resumed with functions to run followed by the argument to run
+-- them with.
+local function runEventHandlerInFreeThread(...)
+	acquireRunnerThreadAndCallEventHandler(...)
+	while true do
+		acquireRunnerThreadAndCallEventHandler(coroutine.yield())
+	end
+end
+
+-- Connection class
+local Connection = {}
+Connection.__index = Connection
+
+function Connection.new(signal, fn)
+	return setmetatable({
+		_connected = true,
+		_signal = signal,
+		_fn = fn,
+		_next = false,
+	}, Connection)
+end
+
+function Connection:Disconnect()
+	assert(self._connected, "Can't disconnect a connection twice.", 2)
+	self._connected = false
+
+	-- Unhook the node, but DON'T clear it. That way any fire calls that are
+	-- currently sitting on this node will be able to iterate forwards off of
+	-- it, but any subsequent fire calls will not hit it, and it will be GCed
+	-- when no more fire calls are sitting on it.
+    local signal = self._signal
+	if signal._handlerListHead == self then
+		signal._handlerListHead = self._next
+	else
+		local prev = signal._handlerListHead
+		while prev and prev._next ~= self do
+			prev = prev._next
+		end
+		if prev then
+			prev._next = self._next
+		end
+	end
+
+    if signal.connectionsChanged then
+		signal.totalConnections -= 1
+        signal.connectionsChanged:Fire(-1)
+    end
+end
+
+-- Make Connection strict
+setmetatable(Connection, {
+	__index = function(tb, key)
+		error(("Attempt to get Connection::%s (not a valid member)"):format(tostring(key)), 2)
+	end,
+	__newindex = function(tb, key, value)
+		error(("Attempt to set Connection::%s (not a valid member)"):format(tostring(key)), 2)
+	end
+})
+
+-- Signal class
 local Signal = {}
 Signal.__index = Signal
-Signal.ClassName = "Signal"
-Signal.totalConnections = 0
 
-
-
--- CONSTRUCTOR
 function Signal.new(createConnectionsChangedSignal)
-	local self = setmetatable({}, Signal)
-	
-	if createConnectionsChangedSignal then
+	local self = setmetatable({
+		_handlerListHead = false,	
+	}, Signal)
+    if createConnectionsChangedSignal then
+        self.totalConnections = 0
 		self.connectionsChanged = Signal.new()
 	end
-
-	self.connections = {}
-	self.totalConnections = 0
-	self.waiting = {}
-	self.totalWaiting = 0
-
-	return self
+    return self
 end
 
-
-
--- METHODS
-function Signal:Fire(...)
-	for _, connection in pairs(self.connections) do
-		connection.Handler(...)
+function Signal:Connect(fn)
+	local connection = Connection.new(self, fn)
+	if self._handlerListHead then
+		connection._next = self._handlerListHead
+		self._handlerListHead = connection
+	else
+		self._handlerListHead = connection
 	end
-	if self.totalWaiting > 0 then
-		local packedArgs = table.pack(...)
-		for waitingId, _ in pairs(self.waiting) do
-			self.waiting[waitingId] = packedArgs
-		end
-	end
-end
-Signal.fire = Signal.Fire
 
-function Signal:Connect(handler)
-	if not (type(handler) == "function") then
-		error(("connect(%s)"):format(typeof(handler)), 2)
-	end
-	
-	local signal = self
-	local connectionId = HttpService:GenerateGUID(false)
-	local connection = {}
-	connection.Connected = true
-	connection.ConnectionId = connectionId
-	connection.Handler = handler
-	self.connections[connectionId] = connection
-
-	function connection:Disconnect()
-		signal.connections[connectionId] = nil
-		connection.Connected = false
-		signal.totalConnections -= 1
-		if signal.connectionsChanged then
-			signal.connectionsChanged:Fire(-1)
-		end
-	end
-	connection.Destroy = connection.Disconnect
-	connection.destroy = connection.Disconnect
-	connection.disconnect = connection.Disconnect
-	self.totalConnections += 1
 	if self.connectionsChanged then
+        self.totalConnections += 1
 		self.connectionsChanged:Fire(1)
 	end
-
 	return connection
 end
-Signal.connect = Signal.Connect
 
-function Signal:Wait()
-	local waitingId = HttpService:GenerateGUID(false)
-	self.waiting[waitingId] = true
-	self.totalWaiting += 1
-	repeat heartbeat:Wait() until self.waiting[waitingId] ~= true
-	self.totalWaiting -= 1
-	local args = self.waiting[waitingId]
-	self.waiting[waitingId] = nil
-	return unpack(args)
-end
-Signal.wait = Signal.Wait
+-- Disconnect all handlers. Since we use a linked list it suffices to clear the
+-- reference to the head handler.
+function Signal:DisconnectAll()
+	self._handlerListHead = false
 
-function Signal:Destroy()
-	if self.bindableEvent then
-		self.bindableEvent:Destroy()
-		self.bindableEvent = nil
-	end
-	if self.connectionsChanged then
+    if self.connectionsChanged then
 		self.connectionsChanged:Fire(-self.totalConnections)
 		self.connectionsChanged:Destroy()
 		self.connectionsChanged = nil
-	end
-	self.totalConnections = 0
-	for connectionId, connection in pairs(self.connections) do
-		self.connections[connectionId] = nil
+		self.totalConnections = 0
 	end
 end
-Signal.destroy = Signal.Destroy
-Signal.Disconnect = Signal.Destroy
-Signal.disconnect = Signal.Destroy
+Signal.Destroy = Signal.DisconnectAll
+Signal.destroy = Signal.DisconnectAll
 
+-- Signal:Fire(...) implemented by running the handler functions on the
+-- coRunnerThread, and any time the resulting thread yielded without returning
+-- to us, that means that it yielded to the Roblox scheduler and has been taken
+-- over by Roblox scheduling, meaning we have to make a new coroutine runner.
+function Signal:Fire(...)
+	local item = self._handlerListHead
+	while item do
+		if item._connected then
+			if not freeRunnerThread then
+				freeRunnerThread = coroutine.create(runEventHandlerInFreeThread)
+			end
+			task.spawn(freeRunnerThread, item._fn, ...)
+		end
+		item = item._next
+	end
+end
+
+-- Implement Signal:Wait() in terms of a temporary connection using
+-- a Signal:Connect() which disconnects itself.
+function Signal:Wait()
+	local waitingCoroutine = coroutine.running()
+	local cn;
+	cn = self:Connect(function(...)
+		cn:Disconnect()
+		task.spawn(waitingCoroutine, ...)
+	end)
+	return coroutine.yield()
+end
 
 
 return Signal